[OF-1441] feat: adding material base class
