+++ Library/include/CSP/Systems/Assets/Material.h
@@ -22,39 +22,15 @@
 
 namespace csp::systems
 {
-class GLTFMaterial;
-}
-
-namespace csp::json
-{
-class JsonSerializer;
-class JsonDeserializer;
-} // namespace csp::json
-
-void ToJson(csp::json::JsonSerializer& Serializer, const csp::systems::GLTFMaterial& Obj);
-void FromJson(const csp::json::JsonDeserializer& Deserializer, csp::systems::GLTFMaterial& Obj);
-
-namespace csp::systems
-{
-
 /// @brief Enum representing the shader type of a material. Currently not in use.
 enum class EShaderType
 {
 	Standard
 };
 
-/// @brief Defines how to alpha value is interpreted
-/// The alpha value is taken from the fourth component of the base color for metallic-roughness material model
-enum class EAlphaMode
-{
-	Opaque,
-	Mask,
-	Blend
-};
-
 /// @ingroup Asset System
-/// @brief Data class which represents a GLTF material.
-class CSP_API GLTFMaterial
+/// @brief Base class for a material.
+class CSP_API Material
 {
 public:
 	/// @brief Sets the user-defined name of the material
@@ -64,133 +40,6 @@ public:
 	/// @brief Gets the user-defined name of the material
 	/// @return csp::common::String&
 	const csp::common::String& GetName() const;
-
-	/// Sets how to alpha value is interpreted
-	/// @param Mode EAlphaMode
-	void SetAlphaMode(EAlphaMode Mode);
-
-	/// Gets how to alpha value is interpreted
-	/// @return EAlphaMode
-	EAlphaMode GetAlphaMode() const;
-
-	/// @brief Sets the alpha cutoff value
-	/// When alphaMode is set to MASK the alphaCutoff property specifies the cutoff threshold.
-	/// If the alpha value is greater than or equal to the alphaCutoff value then it is rendered as fully opaque,
-	/// otherwise, it is rendered as fully transparent. alphaCutoff value is ignored for other modes.
-	/// @param Cutoff float
-	void SetAlphaCutoff(float Cutoff);
-
-	/// Gets the alpha cutoff value
-	/// @return float
-	float GetAlphaCutoff() const;
-
-	/// @brief Sets the doubleSided property which specifies whether the material is double sided.
-	/// When this value is false, back - face culling is enabled, i.e., only front-facing triangles are rendered.
-	/// When this value is true, back - face culling is disabled and double sided lighting is enabled.
-	/// @param DoubleSided bool
-	void SetDoubleSided(bool DoubleSided);
-
-	/// @brief Gets the double sided value
-	/// @return bool
-	bool GetDoubleSided() const;
-
-	/// @brief Sets the factors for the base color of the material.
-	/// This value defines linear multipliers for the sampled texels of the base color texture.
-	/// @param Factor const csp::common::Vector4&
-	void SetBaseColorFactor(const csp::common::Vector4& Factor);
-
-	/// @brief Gets the factor of the base color texture
-	/// @return csp::common::Vector4&
-	const csp::common::Vector4& GetBaseColorFactor() const;
-
-	/// @brief Sets the factor for the metalness of the material.
-	/// This value defines a linear multiplier for the sampled metalness values of the metallic-roughness texture.
-	/// @param Factor float
-	void SetMetallicFactor(float Factor);
-
-	/// @brief Gets the factor of the metallic texture
-	/// @return float
-	float GetMetallicFactor() const;
-
-	/// @brief Sets the factor for the roughness of the material.
-	/// This value defines a linear multiplier for the sampled roughness values of the metallic-roughness texture.
-	/// @param Factor float
-	void SetRoughnessFactor(float Factor);
-
-	/// @brief Gets the factor of the roughness texture
-	/// @return float
-	float GetRoughnessFactor() const;
-
-	/// @brief Sets factors for the emissive color of the material.
-	/// This value defines linear multipliers for the sampled texels of the emissive texture.
-	/// @param Factor const csp::common::Vector3&
-	void SetEmissiveFactor(const csp::common::Vector3& Factor);
-
-	/// @brief Gets the factor of the emissive color texture
-	/// @return csp::common::Vector3&
-	const csp::common::Vector3& GetEmissiveFactor() const;
-
-	/// @brief Sets the base color texture. The first three components (RGB) MUST be encoded with the sRGB transfer function.
-	/// They specify the base color of the material.
-	/// If the fourth component (A) is present, it represents the linear alpha coverage of the material.
-	/// Otherwise, the alpha coverage is equal to 1.0. The material.alphaMode property specifies how alpha is interpreted.
-	/// The stored texels MUST NOT be premultiplied.
-	/// When undefined, the texture MUST be sampled as having 1.0 in all components.
-	/// @param Texture const TextureInfo&
-	void SetBaseColorTexture(const TextureInfo& Texture);
-
-	/// @brief Gets the base color texture
-	/// @return const TextureInfo&
-	const TextureInfo& GetBaseColorTexture() const;
-
-	/// @brief Sets the metallic-roughness texture.
-	/// The metalness values are sampled from the B channel.
-	/// The roughness values are sampled from the G channel.
-	/// These values MUST be encoded with a linear transfer function.
-	/// If other channels are present (R or A), they MUST be ignored for metallic-roughness calculations.
-	/// When undefined, the texture MUST be sampled as having 1.0 in G and B components.
-	/// @param Texture const TextureInfo&
-	void SetMetallicRoughnessTexture(const TextureInfo& Texture);
-
-	/// @brief Gets the metallic-roughness texture
-	/// @return const TextureInfo&
-	const TextureInfo& GetMetallicRoughnessTexture() const;
-
-	/// @brief Sets the tangent space normal texture.
-	/// The texture encodes RGB components with linear transfer function.
-	/// Each texel represents the XYZ components of a normal vector in tangent space.
-	/// The normal vectors use the convention +X is right and +Y is up. +Z points toward the viewer.
-	/// If a fourth component (A) is present, it MUST be ignored. When undefined, the material does not have a tangent space normal texture.
-	/// @param Texture const TextureInfo&
-	void SetNormalTexture(const TextureInfo& Texture);
-
-	/// @brief Gets the tangent space normal texture.
-	/// @return const TextureInfo&
-	const TextureInfo& GetNormalTexture() const;
-
-	/// @brief Sets the occlusion texture.
-	/// The occlusion values are linearly sampled from the R channel.
-	/// Higher values indicate areas that receive full indirect lighting and lower values indicate no indirect lighting.
-	/// If other channels are present (GBA), they MUST be ignored for occlusion calculations.
-	/// When undefined, the material does not have an occlusion texture.
-	/// @param Texture const TextureInfo&
-	void SetOcclusionTexture(const TextureInfo& Texture);
-
-	/// @brief Gets the occlusion texture
-	/// @return const TextureInfo&
-	const TextureInfo& GetOcclusionTexture() const;
-
-	/// @brief Sets the emissive texture.
-	/// It controls the color and intensity of the light being emitted by the material.
-	/// This texture contains RGB components encoded with the sRGB transfer function.
-	/// If a fourth component (A) is present, it MUST be ignored.
-	/// When undefined, the texture MUST be sampled as having 1.0 in RGB components.
-	/// @param Texture const TextureInfo&
-	void SetEmissiveTexture(const TextureInfo& Texture);
-
-	/// @brief Gets the emissive texture
-	/// @return const TextureInfo&
-	const TextureInfo& GetEmissiveTexture() const;
 
 	/// @brief Gets the asset collection id for where this material is stored
 	/// @return const csp::common::String&
@@ -203,34 +52,15 @@ public:
 	/// @brief Constructor which links the material to an asset
 	/// @param AssetCollectionId const csp::common::String& : The asset collection where the material info is stored
 	/// @param AssetId const csp::common::String& : The asset where the material info is stored
-	GLTFMaterial(const csp::common::String& AssetCollectionId, const csp::common::String& AssetId);
+	Material(const csp::common::String& AssetCollectionId, const csp::common::String& AssetId);
+	Material() = default;
 
-private:
+protected:
 	csp::common::String Name;
 	EShaderType Type;
 
-	int Version;
-
-	EAlphaMode AlphaMode;
-	float AlphaCutoff;
-	bool DoubleSided;
-
-	csp::common::Vector4 BaseColorFactor;
-	float MetallicFactor;
-	float RoughnessFactor;
-	csp::common::Vector3 EmissiveFactor;
-
-	TextureInfo BaseColorTexture;
-	TextureInfo MetallicRoughnessTexture;
-	TextureInfo NormalTexture;
-	TextureInfo OcclusionTexture;
-	TextureInfo EmissiveTexture;
-
 	csp::common::String AssetCollectionId;
 	csp::common::String AssetId;
-
-	friend void ::ToJson(csp::json::JsonSerializer& Serializer, const csp::systems::GLTFMaterial& Obj);
-	friend void ::FromJson(const csp::json::JsonDeserializer& Deserializer, csp::systems::GLTFMaterial& Obj);
 };
 
 } // namespace csp::systems
